[bridgeconfig] move bridge flows generation functions to the pkg.
These functions use bridge lock, will convert them to
BridgeConfiguration methods later.

Move test functions related to flow generation.
