+++ go-controller/pkg/node/egress_service_test.go
@@ -19,6 +19,7 @@ import (
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/factory"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/node/controllers/egressservice"
 	nodenft "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/node/nftables"
+	nodetypes "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/node/types"
 	ovntest "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/testing"
 	util "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/util"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/util/mocks"
@@ -299,7 +300,7 @@ var _ = Describe("Egress Service Operations", func() {
 
 				c, err := egressservice.NewController(
 					stopChan,
-					ovnKubeNodeSNATMark,
+					nodetypes.OvnKubeNodeSNATMark,
 					"node",
 					wf.EgressServiceInformer(),
 					wf.ServiceInformer(),
@@ -405,7 +406,7 @@ add element inet ovn-kubernetes egress-service-snat-v4 { 10.128.0.4 comment "nam
 
 				c, err := egressservice.NewController(
 					stopChan,
-					ovnKubeNodeSNATMark,
+					nodetypes.OvnKubeNodeSNATMark,
 					"node",
 					wf.EgressServiceInformer(),
 					wf.ServiceInformer(),
@@ -610,7 +611,7 @@ add element inet ovn-kubernetes egress-service-snat-v4 { 10.128.0.3 comment "nam
 
 				c, err := egressservice.NewController(
 					stopChan,
-					ovnKubeNodeSNATMark,
+					nodetypes.OvnKubeNodeSNATMark,
 					"node",
 					wf.EgressServiceInformer(),
 					wf.ServiceInformer(),
@@ -805,7 +806,7 @@ add element inet ovn-kubernetes egress-service-snat-v4 { 10.128.0.11 comment "na
 
 				c, err := egressservice.NewController(
 					stopChan,
-					ovnKubeNodeSNATMark,
+					nodetypes.OvnKubeNodeSNATMark,
 					"node",
 					wf.EgressServiceInformer(),
 					wf.ServiceInformer(),
@@ -964,7 +965,7 @@ add element inet ovn-kubernetes egress-service-snat-v4 { 10.128.0.11 comment "na
 
 				c, err := egressservice.NewController(
 					stopChan,
-					ovnKubeNodeSNATMark,
+					nodetypes.OvnKubeNodeSNATMark,
 					"node",
 					wf.EgressServiceInformer(),
 					wf.ServiceInformer(),
