+++ go-controller/pkg/node/gateway_udn_test.go
@@ -21,7 +21,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes/fake"
-	utilnet "k8s.io/utils/net"
 	"k8s.io/utils/ptr"
 	"sigs.k8s.io/knftables"
 
@@ -241,105 +240,6 @@ func openflowManagerCheckPorts(ofMgr *openflowManager) {
 		return netConfigs[i].PatchPort < netConfigs[j].PatchPort
 	})
 	Expect(checkPorts(netConfigs, uplink, ofPortPhys)).To(Succeed())
-}
-
-func checkDefaultSvcIsolationOVSFlows(flows []string, defaultConfig *bridgeconfig.BridgeUDNConfiguration, ofPortHost, bridgeMAC string, svcCIDR *net.IPNet) {
-	By(fmt.Sprintf("Checking default service isolation flows for %s", svcCIDR.String()))
-
-	var masqIP string
-	var masqSubnet string
-	var protoPrefix string
-	if utilnet.IsIPv4CIDR(svcCIDR) {
-		protoPrefix = "ip"
-		masqIP = config.Gateway.MasqueradeIPs.V4HostMasqueradeIP.String()
-		masqSubnet = config.Gateway.V4MasqueradeSubnet
-	} else {
-		protoPrefix = "ip6"
-		masqIP = config.Gateway.MasqueradeIPs.V6HostMasqueradeIP.String()
-		masqSubnet = config.Gateway.V6MasqueradeSubnet
-	}
-
-	var nTable0DefaultFlows int
-	var nTable0UDNMasqFlows int
-	var nTable2Flows int
-	for _, flow := range flows {
-		if strings.Contains(flow, fmt.Sprintf("priority=500, in_port=%s, %s, %s_dst=%s, actions=ct(commit,zone=%d,nat(src=%s),table=2)",
-			ofPortHost, protoPrefix, protoPrefix, svcCIDR, config.Default.HostMasqConntrackZone,
-			masqIP)) {
-			nTable0DefaultFlows++
-		} else if strings.Contains(flow, fmt.Sprintf("priority=550, in_port=%s, %s, %s_src=%s, %s_dst=%s, actions=ct(commit,zone=%d,table=2)",
-			ofPortHost, protoPrefix, protoPrefix, masqSubnet, protoPrefix, svcCIDR, config.Default.HostMasqConntrackZone)) {
-			nTable0UDNMasqFlows++
-		} else if strings.Contains(flow, fmt.Sprintf("priority=100, table=2, actions=set_field:%s->eth_dst,output:%s",
-			bridgeMAC, defaultConfig.OfPortPatch)) {
-			nTable2Flows++
-		}
-	}
-
-	Expect(nTable0DefaultFlows).To(Equal(1))
-	Expect(nTable0UDNMasqFlows).To(Equal(1))
-	Expect(nTable2Flows).To(Equal(1))
-}
-
-func checkAdvertisedUDNSvcIsolationOVSFlows(flows []string, netConfig *bridgeconfig.BridgeUDNConfiguration, netName string, svcCIDR *net.IPNet, expectedNFlows int) {
-	By(fmt.Sprintf("Checking advertsised UDN %s service isolation flows for %s; expected %d flows",
-		netName, svcCIDR.String(), expectedNFlows))
-
-	var matchingIPFamilySubnet *net.IPNet
-	var protoPrefix string
-	var udnAdvertisedSubnets []*net.IPNet
-	var err error
-	for _, clusterEntry := range netConfig.Subnets {
-		udnAdvertisedSubnets = append(udnAdvertisedSubnets, clusterEntry.CIDR)
-	}
-	if utilnet.IsIPv4CIDR(svcCIDR) {
-		matchingIPFamilySubnet, err = util.MatchFirstIPNetFamily(false, udnAdvertisedSubnets)
-		Expect(err).ToNot(HaveOccurred())
-		protoPrefix = "ip"
-	} else {
-		matchingIPFamilySubnet, err = util.MatchFirstIPNetFamily(false, udnAdvertisedSubnets)
-		Expect(err).ToNot(HaveOccurred())
-		protoPrefix = "ip6"
-	}
-
-	var nFlows int
-	for _, flow := range flows {
-		if strings.Contains(flow, fmt.Sprintf("priority=200, table=2, %s, %s_src=%s, actions=drop",
-			protoPrefix, protoPrefix, matchingIPFamilySubnet)) {
-			nFlows++
-		}
-		if strings.Contains(flow, fmt.Sprintf("priority=550, in_port=LOCAL, %s, %s_src=%s, %s_dst=%s, actions=ct(commit,zone=64001,table=2)",
-			protoPrefix, protoPrefix, matchingIPFamilySubnet, protoPrefix, svcCIDR)) {
-			nFlows++
-		}
-	}
-
-	Expect(nFlows).To(Equal(expectedNFlows))
-}
-
-func checkUDNSvcIsolationOVSFlows(flows []string, netConfig *bridgeconfig.BridgeUDNConfiguration, netName string, svcCIDR *net.IPNet, expectedNFlows int) {
-	By(fmt.Sprintf("Checking UDN %s service isolation flows for %s; expected %d flows",
-		netName, svcCIDR.String(), expectedNFlows))
-
-	var mgmtMasqIP string
-	var protoPrefix string
-	if utilnet.IsIPv4CIDR(svcCIDR) {
-		mgmtMasqIP = netConfig.V4MasqIPs.ManagementPort.IP.String()
-		protoPrefix = "ip"
-	} else {
-		mgmtMasqIP = netConfig.V6MasqIPs.ManagementPort.IP.String()
-		protoPrefix = "ip6"
-	}
-
-	var nFlows int
-	for _, flow := range flows {
-		if strings.Contains(flow, fmt.Sprintf("priority=200, table=2, %s, %s_src=%s, actions=drop",
-			protoPrefix, protoPrefix, mgmtMasqIP)) {
-			nFlows++
-		}
-	}
-
-	Expect(nFlows).To(Equal(expectedNFlows))
 }
 
 func getDummyOpenflowManager() *openflowManager {
@@ -792,10 +692,10 @@ var _ = Describe("UserDefinedNetworkGateway", func() {
 
 			for _, svcCIDR := range config.Kubernetes.ServiceCIDRs {
 				// Check flows for default network service CIDR.
-				checkDefaultSvcIsolationOVSFlows(flowMap["DEFAULT"], defaultUdnConfig, ofPortHost, bridgeMAC, svcCIDR)
+				bridgeconfig.CheckDefaultSvcIsolationOVSFlows(flowMap["DEFAULT"], defaultUdnConfig, ofPortHost, bridgeMAC, svcCIDR)
 
 				// Expect exactly one flow per UDN for table 2 for service isolation.
-				checkUDNSvcIsolationOVSFlows(flowMap["DEFAULT"], bridgeUdnConfig, "bluenet", svcCIDR, 1)
+				bridgeconfig.CheckUDNSvcIsolationOVSFlows(flowMap["DEFAULT"], bridgeUdnConfig, "bluenet", svcCIDR, 1)
 			}
 
 			// The second call to checkPorts() will return no ofPort for the UDN - simulating a deletion that already was
@@ -822,10 +722,10 @@ var _ = Describe("UserDefinedNetworkGateway", func() {
 
 			for _, svcCIDR := range config.Kubernetes.ServiceCIDRs {
 				// Check flows for default network service CIDR.
-				checkDefaultSvcIsolationOVSFlows(flowMap["DEFAULT"], defaultUdnConfig, ofPortHost, bridgeMAC, svcCIDR)
+				bridgeconfig.CheckDefaultSvcIsolationOVSFlows(flowMap["DEFAULT"], defaultUdnConfig, ofPortHost, bridgeMAC, svcCIDR)
 
 				// Expect no more flows per UDN for table 2 for service isolation.
-				checkUDNSvcIsolationOVSFlows(flowMap["DEFAULT"], bridgeUdnConfig, "bluenet", svcCIDR, 0)
+				bridgeconfig.CheckUDNSvcIsolationOVSFlows(flowMap["DEFAULT"], bridgeUdnConfig, "bluenet", svcCIDR, 0)
 			}
 			return nil
 		})
@@ -1023,10 +923,10 @@ var _ = Describe("UserDefinedNetworkGateway", func() {
 
 			for _, svcCIDR := range config.Kubernetes.ServiceCIDRs {
 				// Check flows for default network service CIDR.
-				checkDefaultSvcIsolationOVSFlows(flowMap["DEFAULT"], defaultUdnConfig, ofPortHost, bridgeMAC, svcCIDR)
+				bridgeconfig.CheckDefaultSvcIsolationOVSFlows(flowMap["DEFAULT"], defaultUdnConfig, ofPortHost, bridgeMAC, svcCIDR)
 
 				// Expect exactly one flow per UDN for tables 0 and 2 for service isolation.
-				checkUDNSvcIsolationOVSFlows(flowMap["DEFAULT"], bridgeUdnConfig, "bluenet", svcCIDR, 1)
+				bridgeconfig.CheckUDNSvcIsolationOVSFlows(flowMap["DEFAULT"], bridgeUdnConfig, "bluenet", svcCIDR, 1)
 			}
 
 			// The second call to checkPorts() will return no ofPort for the UDN - simulating a deletion that already was
@@ -1053,10 +953,10 @@ var _ = Describe("UserDefinedNetworkGateway", func() {
 
 			for _, svcCIDR := range config.Kubernetes.ServiceCIDRs {
 				// Check flows for default network service CIDR.
-				checkDefaultSvcIsolationOVSFlows(flowMap["DEFAULT"], defaultUdnConfig, ofPortHost, bridgeMAC, svcCIDR)
+				bridgeconfig.CheckDefaultSvcIsolationOVSFlows(flowMap["DEFAULT"], defaultUdnConfig, ofPortHost, bridgeMAC, svcCIDR)
 
 				// Expect no more flows per UDN for tables 0 and 2 for service isolation.
-				checkUDNSvcIsolationOVSFlows(flowMap["DEFAULT"], bridgeUdnConfig, "bluenet", svcCIDR, 0)
+				bridgeconfig.CheckUDNSvcIsolationOVSFlows(flowMap["DEFAULT"], bridgeUdnConfig, "bluenet", svcCIDR, 0)
 			}
 			return nil
 		})
@@ -1264,10 +1164,10 @@ var _ = Describe("UserDefinedNetworkGateway", func() {
 
 			for _, svcCIDR := range config.Kubernetes.ServiceCIDRs {
 				// Check flows for default network service CIDR.
-				checkDefaultSvcIsolationOVSFlows(flowMap["DEFAULT"], defaultUdnConfig, ofPortHost, bridgeMAC, svcCIDR)
+				bridgeconfig.CheckDefaultSvcIsolationOVSFlows(flowMap["DEFAULT"], defaultUdnConfig, ofPortHost, bridgeMAC, svcCIDR)
 
 				// Expect exactly one flow per advertised UDN for table 2 and table 0 for service isolation.
-				checkAdvertisedUDNSvcIsolationOVSFlows(flowMap["DEFAULT"], bridgeUdnConfig, "bluenet", svcCIDR, 2)
+				bridgeconfig.CheckAdvertisedUDNSvcIsolationOVSFlows(flowMap["DEFAULT"], bridgeUdnConfig, "bluenet", svcCIDR, 2)
 			}
 
 			// The second call to checkPorts() will return no ofPort for the UDN - simulating a deletion that already was
@@ -1294,10 +1194,10 @@ var _ = Describe("UserDefinedNetworkGateway", func() {
 
 			for _, svcCIDR := range config.Kubernetes.ServiceCIDRs {
 				// Check flows for default network service CIDR.
-				checkDefaultSvcIsolationOVSFlows(flowMap["DEFAULT"], defaultUdnConfig, ofPortHost, bridgeMAC, svcCIDR)
+				bridgeconfig.CheckDefaultSvcIsolationOVSFlows(flowMap["DEFAULT"], defaultUdnConfig, ofPortHost, bridgeMAC, svcCIDR)
 
 				// Expect no more flows per UDN for table 2 and table0 for service isolation.
-				checkAdvertisedUDNSvcIsolationOVSFlows(flowMap["DEFAULT"], bridgeUdnConfig, "bluenet", svcCIDR, 0)
+				bridgeconfig.CheckAdvertisedUDNSvcIsolationOVSFlows(flowMap["DEFAULT"], bridgeUdnConfig, "bluenet", svcCIDR, 0)
 			}
 			return nil
 		})
