+++ go-controller/pkg/node/openflow_manager.go
@@ -14,6 +14,7 @@ import (
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/config"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/generator/udn"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/node/bridgeconfig"
+	nodetypes "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/node/types"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/types"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/util"
 )
@@ -210,12 +211,12 @@ func (c *openflowManager) updateBridgePMTUDFlowCache(key string, ipAddrs []strin
 	c.defaultBridge.Mutex.Lock()
 	defer c.defaultBridge.Mutex.Unlock()
 
-	dftFlows := pmtudDropFlows(c.defaultBridge, ipAddrs)
+	dftFlows := bridgeconfig.PmtudDropFlows(c.defaultBridge, ipAddrs)
 	c.updateFlowCacheEntry(key, dftFlows)
 	if c.externalGatewayBridge != nil {
 		c.externalGatewayBridge.Mutex.Lock()
 		defer c.externalGatewayBridge.Mutex.Unlock()
-		exGWBridgeDftFlows := pmtudDropFlows(c.externalGatewayBridge, ipAddrs)
+		exGWBridgeDftFlows := bridgeconfig.PmtudDropFlows(c.externalGatewayBridge, ipAddrs)
 		c.updateExBridgeFlowCacheEntry(key, exGWBridgeDftFlows)
 	}
 }
@@ -230,11 +231,11 @@ func (c *openflowManager) updateBridgeFlowCache(hostIPs []net.IP, hostSubnets []
 	// CAUTION: when adding new flows where the in_port is ofPortPatch and the out_port is ofPortPhys, ensure
 	// that dl_src is included in match criteria!
 
-	dftFlows, err := flowsForDefaultBridge(c.defaultBridge, hostIPs)
+	dftFlows, err := bridgeconfig.FlowsForDefaultBridge(c.defaultBridge, hostIPs)
 	if err != nil {
 		return err
 	}
-	dftCommonFlows, err := commonFlows(hostSubnets, c.defaultBridge)
+	dftCommonFlows, err := bridgeconfig.CommonFlows(hostSubnets, c.defaultBridge)
 	if err != nil {
 		return err
 	}
@@ -248,7 +249,7 @@ func (c *openflowManager) updateBridgeFlowCache(hostIPs []net.IP, hostSubnets []
 		c.externalGatewayBridge.Mutex.Lock()
 		defer c.externalGatewayBridge.Mutex.Unlock()
 		c.updateExBridgeFlowCacheEntry("NORMAL", []string{fmt.Sprintf("table=0,priority=0,actions=%s\n", util.NormalAction)})
-		exGWBridgeDftFlows, err := commonFlows(hostSubnets, c.externalGatewayBridge)
+		exGWBridgeDftFlows, err := bridgeconfig.CommonFlows(hostSubnets, c.externalGatewayBridge)
 		if err != nil {
 			return err
 		}
@@ -357,10 +358,10 @@ func bootstrapOVSFlows(nodeName string) error {
 	// for non-IP packets that would normally be forwarded with NORMAL action (table 0, priority 0 flow).
 	dftFlows = append(dftFlows,
 		fmt.Sprintf("cookie=%s, priority=10, table=0, in_port=%s, dl_src=%s, actions=output:NORMAL",
-			defaultOpenFlowCookie, ofportPatch, bridgeMACAddress))
+			nodetypes.DefaultOpenFlowCookie, ofportPatch, bridgeMACAddress))
 	dftFlows = append(dftFlows,
 		fmt.Sprintf("cookie=%s, priority=9, table=0, in_port=%s, actions=drop",
-			defaultOpenFlowCookie, ofportPatch))
+			nodetypes.DefaultOpenFlowCookie, ofportPatch))
 	dftFlows = append(dftFlows, "priority=0, table=0, actions=output:NORMAL")
 
 	_, stderr, err = util.ReplaceOFFlows(bridge, dftFlows)
